+++ google/index.ts
@@ -1,4 +1,10 @@
-import { resolveToken, stringizeScope } from './helpers';
+import {
+  resolveToken,
+  stringizeScope,
+  constructFields,
+  createJwtToken,
+  googleApiRequest,
+} from './helpers';
 
 export { api as oauthAuthenticationApis } from './helpers';
 export { api as oauthRouteApis } from './routes';
@@ -9,4 +15,7 @@ export { default as serviceAccountHandler } from './handlers/sac';
 export const oauthHelpers = {
   resolveToken,
   stringizeScope,
+  constructFields,
+  createJwtToken,
+  googleApiRequest,
 };